fix(permissions): hide module-status validate button from standard users (#1411)
Gate the sidebar Validate/Unvalidate button on a dedicated custom
permission `module.status/<cf>` (cost center), granted to unit-breadth
users (principal) only — standard (own-scope) users never receive it, so
the button is hidden. Frontend gates UI on permissions, never roles; the
PATCH stays enforced by `require_module_unit_scope`.

Also consolidate the frontend permission surface:
- Delete `constant/permissions.ts`; fold the enum/types + MODULE_STATUS_PERMISSION
  into the store-free `utils/permission.ts` leaf (keeps pure unit tests runnable).
- Make `stores/auth.ts` the single entry point (re-exports PermissionAction);
  all consumers call `authStore.*`.
- Collapse the three router guards into one `permissionGuard`; delete the unused
  `requirePermission` factory and the dead `meta.roles`/roles_raw branch in
  authGuard (zero routes used it — the app's only role check).
- Remove now-orphaned `hasUnitScopePermission` / `hasUserUnitScopePermission`
  (scope-inference superseded by the dedicated `module.status` key) and its tests.

Backend grant + tests; docs updated (06-PERMISSION-SYSTEM, frontend overview).
